+++ src/comm_select.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: comm_select.cc,v 1.78 2007/04/28 22:26:37 hno Exp $
+ * $Id: comm_select.cc,v 1.79 2007/04/30 16:56:09 wessels Exp $
  *
  * DEBUG: section 5     Socket Functions
  *
@@ -687,7 +687,7 @@ comm_select(int msec)
     } while (timeout > current_dtime)
 
         ;
-    debugs(5, 8, "comm_select: time out: " << (int) squid_curtime);
+    debugs(5, 8, "comm_select: time out: " << squid_curtime);
 
     return COMM_TIMEOUT;
 }
